==> notebook/01.create_final_dataset.ipynb <==
# Create final dataset

import pandas as pd

df_payements = pd.read_csv('payments_amount_usd.csv')
df_payements.dtypes

## Add month

df_payements['date'] = pd.to_datetime(df_payements['date'])
df_payements['month'] = df_payements['date'].dt.to_period('M')

## Merge with pledges

df_pledges = pd.read_json('pledges.json')
df_pledges.dtypes

df_merged = df_payements.merge(df_pledges, how='inner', on='pledge_id', suffixes=('_payements', '_pledges'))

df_merged.dtypes

cols_to_drop = [col for col in df_merged.columns if col.endswith('_pledges')]

cols_to_drop

df_merged = df_merged.drop(columns=cols_to_drop)

df_merged.dtypes

cols_to_rename = [col for col in df_merged.columns if col.endswith('_payements')]
cols_to_rename

df_merged.rename(
    columns=lambda col: col.replace('_payements', '') if col.endswith('_payements') else col,
    inplace=True
)

df_merged.head()

df_merged.dtypes

- Drop unnamed

df_merged.dtypes

## Export to CSV

df_merged.shape
df_merged.memory_usage(deep=True).sum() / (1024**2)  # en Mo

df_merged.dtypes

df_merged = df_merged.drop(columns='Unnamed: 0')

df_merged.to_csv('payments_and_pledges.csv', index=False)
